Bogota map added

+++ src/ds/graph.rs
@@ -1,7 +1,7 @@
 #[derive(Debug)]
 pub struct Graph {
     pub edges: Vec<Vec<usize>>,
-    pub weights: Vec<Vec<u32>>,
+    pub weights: Vec<Vec<f64>>,
 }
 
 impl Graph {
@@ -12,7 +12,7 @@ impl Graph {
         }
     }
 
-    pub fn add_edge(&mut self, u: usize, v: usize, w: u32) {
+    pub fn add_edge(&mut self, u: usize, v: usize, w: f64) {
         self.edges[u].push(v);
         self.weights[u].push(w);
     }
